feat: modernizar API con Spring Boot 4, validación, MapStruct y tests
- Java 25, validación Jakarta, Actuator y perfiles dev/prod/test

- BigDecimal, excepciones globales, paginación y MapStruct

- Suite de tests (unitarios, repositorio, MVC, integración, mappers)

- CI con GitHub Actions, JaCoCo, README y .env.example

Made-with: Cursor

+++ src/main/java/com/example/demo/model/Producto.java
@@ -6,17 +6,21 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.*;
 
-@Getter @Setter
+import java.math.BigDecimal;
+
+@Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Entity
 public class Producto {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String nombre;
     private String categoria;
-    private Double precio;
+    private BigDecimal precio;
     private int cantidad;
 }
